requirements.txt
paddlepaddle-gpu
paddleocr
pyyaml
pymupdf
numpy
opencv-python
rapidocr-onnxruntime
